chore: add benchmark to ftm script

+++ packages/yswaps/contracts/interfaces/ISolidlyRouter.sol
@@ -3,28 +3,28 @@ pragma solidity >=0.8.4 <0.9.0;
 
 interface ISolidlyRouter {
   struct route {
-        address from;
-        address to;
-        bool stable;
-    }
+    address from;
+    address to;
+    bool stable;
+  }
 
-  function getAmountsOut(uint, route[] memory) external view returns (uint[] memory amounts);
+  function getAmountsOut(uint256, route[] memory) external view returns (uint256[] memory amounts);
 
   function swapExactTokensForTokensSimple(
-        uint,
-        uint,
-        address,
-        address,
-        bool,
-        address,
-        uint
-    ) external returns (uint[] memory);
+    uint256,
+    uint256,
+    address,
+    address,
+    bool,
+    address,
+    uint256
+  ) external returns (uint256[] memory);
 
   function swapExactTokensForTokens(
-        uint,
-        uint,
-        route[] calldata,
-        address,
-        uint
-    ) external returns (uint[] memory amounts);
+    uint256,
+    uint256,
+    route[] calldata,
+    address,
+    uint256
+  ) external returns (uint256[] memory amounts);
 }
